Move getReviews to utils.js
